refactor(analysis-types): split source DTO
