+++ crates/tokmd-analysis-types/src/lib.rs
@@ -30,6 +30,7 @@ mod fun;
 mod git;
 mod imports;
 mod license;
+mod source;
 mod topics;
 pub mod util;
 
@@ -72,6 +73,7 @@ pub use git::{
 };
 pub use imports::{ImportEdge, ImportReport};
 pub use license::{LicenseFinding, LicenseReport, LicenseSourceKind};
+pub use source::AnalysisSource;
 pub use topics::{TopicClouds, TopicTerm};
 pub use util::{
     AnalysisLimits, empty_file_row, is_infra_lang, is_test_path, normalize_path, normalize_root,
@@ -113,19 +115,6 @@ pub struct AnalysisReceipt {
     pub api_surface: Option<ApiSurfaceReport>,
     pub effort: Option<EffortEstimateReport>,
     pub fun: Option<FunReport>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AnalysisSource {
-    pub inputs: Vec<String>,
-    pub export_path: Option<String>,
-    pub base_receipt_path: Option<String>,
-    pub export_schema_version: Option<u32>,
-    pub export_generated_at_ms: Option<u128>,
-    pub base_signature: Option<String>,
-    pub module_roots: Vec<String>,
-    pub module_depth: usize,
-    pub children: String,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
