+++ crates/tokmd-analysis-types/src/source.rs
@@ -1,0 +1,15 @@
+use serde::{Deserialize, Serialize};
+
+/// Source metadata recorded in an analysis receipt.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AnalysisSource {
+    pub inputs: Vec<String>,
+    pub export_path: Option<String>,
+    pub base_receipt_path: Option<String>,
+    pub export_schema_version: Option<u32>,
+    pub export_generated_at_ms: Option<u128>,
+    pub base_signature: Option<String>,
+    pub module_roots: Vec<String>,
+    pub module_depth: usize,
+    pub children: String,
+}
